\name{ChoosePTF}
\alias{ChoosePTF}
\title{
Choosing the right PTF
}
\description{
Launches a graphical interface helping to choose the right PTFs depending on the user available input data and desired output.
}
\usage{
ChoosePTF()
}
\details{
Calling ChoosePTF opens a guiWidget window. The available data have to be selected, then the 'What PTF?' button must be clicked on to show the PTFs to be used for each output.
}
\value{
Returns a guiWidget of type: gWindowtcltk for toolkit: guiWidgetsToolkittcltk 
}
\references{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}

\seealso{
\code{\link{predict.ptf}}
}
