\name{PTF01}
\alias{PTF01}
\docType{data}
\title{
PTF01}
\description{
PTF01 is a regression tree based PTF predicting the saturated water content THS from modified FAO texture class (see \code{\link{psd2classFAO_MOD}}) and top/subsoil distinction (T/S).}
\usage{data(PTF01)}
\format{
  \code{\link[mvpart:rpart.object]{rpart object}}.
}
\details{
PTF01 was built with \code{\link[mvpart:rpart]{mvpart:rpart}.
}}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\source{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
