\name{PTF03}
\alias{PTF03}
\docType{data}
\title{
PTF03}
\description{
PTF03 is a regression tree based PTF predicting the saturated water content THS from USDA texture class (see \code{\link{psd2classUS}}) and top/subsoil distinction (T/S).}
\usage{data(PTF03)}
\format{
  \code{\link[mvpart:rpart.object]{rpart object}}.
}
\details{
PTF03 was built with \code{\link[mvpart:rpart]{mvpart:rpart}.
}}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\source{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
