\name{PTF04}
\alias{PTF04}
\docType{data}
\title{
PTF04}
\description{
PTF04 is a regression tree based PTF predicting the saturated water content THS from particle size distribution (PSD: sand, silt, clay percentages), top/subsoil distinction (T/S) and organic carbon content (OC, mass percentage).}
\usage{data(PTF04)}
\format{
  \code{\link[mvpart:rpart.object]{rpart object}}.
}
\details{
PTF04 was built with \code{\link[mvpart:rpart]{mvpart:rpart}.
}}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\source{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
