\name{PTF06}
\alias{PTF06}
\docType{data}
\title{
PTF06}
\description{
is a linear regression based PTF predicting the saturated water content (THS) from  PSD, T/S, BD and pH. For the format and the details, see \code{\link{PTF05}}.
}
\usage{data(PTF06)}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\references{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
