\name{PTF09}
\alias{PTF09}
\docType{data}
\title{
PTF09}
\description{
PTF09 is a linear regression based PTF predicting the water content at field capacity (FC) from  PSD and OC. For the format and details, see \code{\link{PTF05}}.}
\usage{data(PTF09)}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\references{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
