\name{PTF12}
\alias{PTF12}
\docType{data}
\title{
PTF12}
\description{
PTF12 is a linear regression based PTF predicting the water content at wilting point (WP) from  PSD and OC. For the format, see \code{\link{PTF05}}.}
\usage{data(PTF12)}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\references{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
