\name{PTF13}
\alias{PTF13}
\docType{data}
\title{
PTF13}
\description{
PTF13 is a regression tree based PTF predicting the saturated hydraulic conductivity (log10(KS)) from modified FAO texture class and T/S. For the format and details, see \code{\link{PTF04}}.}
\usage{data(PTF13)}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\references{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
