\name{PTF15}
\alias{PTF15}
\docType{data}
\title{
PTF15}
\description{
PTF15 is a regression tree based PTF predicting the saturated hydraulic conductivity (log10(KS)) from USDA texture class and T/S. For the format and details, see \code{\link{PTF04}}.}
\usage{data(PTF15)}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\source{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
