\name{PTF17}
\alias{PTF17}
\docType{data}
\title{
PTF17}
\description{
PTF17 is a linear regression based PTF predicting the saturated hydraulic conductivity (log10(KS)) from PSD, T/S, pH and CEC. For the format, see \code{\link{PTF05}}.}
\usage{data(PTF17)}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\source{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
