\name{PTF21}
\alias{PTF21}
\docType{data}
\title{
PTF21}
\description{
PTF21 is a linear regression base PTF predicting the parameters of van Genuchten model from PSD, T/S, OC and BD. For the format, see \code{\link{PTF20}}.}
\usage{data(PTF21)}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\source{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
