\name{PTF22}
\alias{PTF22}
\docType{data}
\title{
PTF22}
\description{
PTF22 is a linear regression base PTF predicting the parameters of van Genuchten model from PSD, T/S, OC, BD and pH. For the format, see \code{\link{PTF20}}.}
\usage{data(PTF22)}
\seealso{
\code{\link{predict.ptf}}, \code{\link{ChoosePTF}, \code{\link{ptf.data}}}
}
\source{
Tóth Brigitta, Weynants Mélanie, Nemes Attila, Makó Andras, Bilas George and Tóth Gergely. 2014. New generation of hydraulic pedotransfer functions for Europe. European Journal of Soil Science. in Press.
}
\keyword{datasets}
